requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
django-extensions==3.2.3
django-form-utils==1.0.3
django-moneyfield==0.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
openpyxl==3.1.2
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
